Added rating calculation

+++ models/learnlist.py
@@ -4,7 +4,7 @@ from models.dictionary import Dictionary
 class LearnList(db.Model):
     twitter_user       = db.StringProperty()
     dict_entry         = db.ReferenceProperty(Dictionary)
-    interval_days      = db.IntegerProperty()
+    interval_days      = db.FloatProperty()
     efactor            = db.FloatProperty()
     next_serve_time    = db.IntegerProperty()
     next_serve_date    = db.DateProperty()
